update calc + tests

+++ src/main/java/AllOperations/Operation.java
@@ -2,20 +2,15 @@ package AllOperations;
 
 public abstract class Operation {
 
-
     private final char operatorSymbol;
 
     public Operation(char operatorSymbol) {
         this.operatorSymbol = operatorSymbol;
     }
 
-
     public char getOperatorSymbol() {
         return operatorSymbol;
     }
 
-
-    public abstract int doOperation(int first, int second) throws Exception;
-
-
+    public abstract double doOperation(double first, double second) throws Exception;
 }
